Tune fuse.js sort function.
Original was sorting by index when the weights were:

  0.003 and 2.2E-16.

The 2.2E-16 was sorting after the 0.003 because they fell into the
same bucket and 0.003 was earlier in the commands array.

Added a minimum size that score must be larger than to sort by order.
Also added a clause that checks the small scores and if they are the
same, only then does it sort by index.